% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment2palette.R
\name{augment2palette}
\alias{augment2palette}
\title{Assistance in adding a new palette to the package}
\usage{
augment2palette(augment, palette.name.new)
}
\arguments{
\item{augment}{The finalized augment dataframe, presumed to have been built iteratively with \code{palette_augment_helper()}.
Data frame with taxa (or other identifier names) in column "name" and colors in column "color"}

\item{palette.name.new}{Name for the new palette. Ensure this name is not already used for a different palette.}
}
\value{
Nothing; prints instructions to console.
}
\description{
This function is based on the assumption that you'll develop/check a new palette using \code{palette_augment_helper()}.
Here the function gives instructions for how to convert your finalized augment dataframe into its own palette.
Note that this function does NOT change anything in the package, merely gives instructions on how to make those changes.
}
\details{
This package is written so that the main changes needed to include a new palette are
all in the function \code{make_colorvec()}. This includes adding in a hard-coded vector for the new palette,
updating the list of acceptable palettes within the function (\code{palette.name}) to make sure
that the new palette name gets through the error-checker, and updating the documentation of that file.

I also recommend updating Readme.Rmd to describe the new palette and provide the color scheme
with \code{palette_vis()}. While not strictly necessary, this will help labmates and other package
users find the details without reading through help files.

This package template was built using the tidyverse package development workflow;
adding to it or modifying it will be far easier if you also use that workflow. See \href{https://r-pkgs.org/whole-game.html}{this chapter} for an overview.
}
\examples{
aug = data.frame(name = c("asclepias", "drosophila"),
color = c("coral", "purple"))
augment2palette(augment = aug,
               palette.name.new = "testPalette")
}
